changed to double

+++ src/main/model/Expense.java
@@ -5,14 +5,14 @@ import persistence.Writable;
 
 //Class that makes an expense contained in a budget
 public class Expense implements Writable {
-    private Integer cost;
+    private Double cost;
     private String name;
     private String category;
 
     //Requires: expense to be an integer
     //Modifies: this
     //effects: creates new expense
-    public Expense(String name, Integer cost, String category) {
+    public Expense(String name, Double cost, String category) {
         this.name = name;
         this.cost = cost;
         this.category = category;
@@ -41,7 +41,7 @@ public class Expense implements Writable {
         return name;
     }
 
-    public Integer getCost() {
+    public Double getCost() {
         return cost;
     }
 }
